+++ partial/dev/dev-spec.js
@@ -1,0 +1,18 @@
+describe('DevController', function() {
+
+	beforeEach(module('generic.table'));
+
+	var scope,ctrl;
+
+    beforeEach(inject(function($rootScope, $controller) {
+      scope = $rootScope.$new();
+      ctrl = $controller('DevController', {$scope: scope});
+    }));
+
+	it('should ...', inject(function() {
+
+		expect(1).toEqual(1);
+
+	}));
+
+});