+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 Release History
--------------
+---------------
+* 22/4/2016  - v1.0.2 Added support for exporting table data to CSV, fix for column sort click function
+* 20/4/2016  - v1.0.1 Added support for custom click function for column, improved mapping function
 * 19/4/2016  - v1.0.0 Initial release of generic-tables with support for pagination, sorting, custom render function and basic events